Moved backend errors to another file
